WIP
A few months ago I've started a hure refactoring of the project to make
a further step from first prototype to actual application. But then I
run out of time. A couple of times I sit again on the project doing some
refactoring, but still didn't got a chance to finish it :( Commiting it
as is in order to open source what I have so far.

Currently it discovers sockets and without connecting to a socket reads
its status (enabled/disabled) from advertised data packet.